Move $GetPartial into get to locate it next to the other Get logic
